itcl-quantization-toolkit/model_converters/tflite2json/pyproject.toml
[tool.poetry]
name = "itcl-tflite2json"
version = "0.1.5"
description = ""
authors = ["Jorge Ruiz Gómez <jorge.ruiz@itcl.es>"]
packages = [
    {include = "itcl_tflite2json", from="."}
]

[tool.poetry.dependencies]
python = "~3.10"
tensorflow = "^2.9.1"
itcl-quantization-toolkit = { path = "../../core/", develop = true }
tflite = "^2.4.0"
PyYAML = "^6.0"


[tool.poetry.dev-dependencies]
ipykernel = "^6.15.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
